+++ Tools/AdminToolbox.ps1
@@ -27,7 +27,7 @@ Function out-manifests {
         HelpInfoUri          = "https://github.com/TheTaylorLee/AdminToolbox/issues"
         IconUri              = 'https://avatars2.githubusercontent.com/u/53202926?s=460&v=4'
         LicenseUri           = 'https://github.com/TheTaylorLee/AdminToolbox/blob/master/LICENSE.txt'
-        ModuleVersion        = "9.1.0"
+        ModuleVersion        = "9.2.0"
         Powershellversion    = "5.1"
         ProjectUri           = 'https://github.com/TheTaylorLee/AdminToolbox'
         RequiredModules      = (
@@ -36,7 +36,7 @@ Function out-manifests {
             @{ModuleName = 'AdminToolbox.Exchange'; ModuleVersion = '1.8.0'; },
             @{ModuleName = 'AdminToolbox.FFTools'; ModuleVersion = '4.7.0'; },
             @{ModuleName = 'AdminToolbox.FileManagement'; ModuleVersion = '1.7.0'; },
-            @{ModuleName = 'AdminToolbox.FortiWizard'; ModuleVersion = '2.0.0'; },
+            @{ModuleName = 'AdminToolbox.FortiWizard'; ModuleVersion = '2.1.0'; },
             @{ModuleName = 'AdminToolbox.Fun'; ModuleVersion = '1.5.0'; },
             @{ModuleName = 'AdminToolbox.Networking'; ModuleVersion = '2.9.0'; },
             @{ModuleName = 'AdminToolbox.Office365'; ModuleVersion = '2.5.0'; },
@@ -178,11 +178,11 @@ Function out-manifests {
     $savepath = "$env:gitrootpath\AdminToolbox.FortiWizard"
     $Params = @{
         CompatiblePSEditions = "Desktop", "Core"
-        FunctionsToExport    = 'Disable-SipALG', 'Enable-Management', 'Enable-PushNotifications', 'Get-FortiWizard', 'New-AddressObject', 'New-AddressGroup', 'Write-SSLVPNConfig', 'New-IPPoolOverload', 'New-IPPoolFixedRange', 'New-VIPHost', 'New-VIPRange', 'New-ServiceObject', 'New-ServiceGroup', 'New-P2PTunnel', 'Show-Docs', 'New-P2PTunnelNAT', 'New-DialUPTunnelRemoteNAT', 'New-DialUPTunnelBehindNAT'
+        FunctionsToExport    = 'Disable-SipALG', 'Enable-Management', 'Enable-PushNotifications', 'Get-FortiWizard', 'New-AddressObject', 'New-AddressGroup', 'Write-SSLVPNConfig', 'New-IPPoolOverload', 'New-IPPoolFixedRange', 'New-VIPHost', 'New-VIPRange', 'New-ServiceObject', 'New-ServiceGroup', 'New-P2PTunnel', 'Show-Docs', 'New-P2PTunnelNAT', 'New-DialUPTunnelRemoteNAT', 'New-DialUPTunnelBehindNAT', 'New-FormTunnel'
         Path                 = "$savepath\AdminToolbox.FortiWizard.psd1"
         Author               = "Taylor Lee"
         Description          = "Functions that generate configuration scripts and manage FortiOS. This Module is unofficial and in no way affiliated with Fortinet."
-        ModuleVersion        = "2.0.0"
+        ModuleVersion        = "2.1.0"
         Powershellversion    = "5.1"
         RootModule           = "AdminToolbox.FortiWizardManifest.psm1"
         ReleaseNotes         = "Full ChangeLog contained in bundled ChangeLog"